src/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.24.89
botocore==1.27.89
Django==4.0.3
django-cors-headers==3.11.0
django-storages==1.13.1
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.1
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12